feat(scaffold): replace flat material volume with sparse sub-block packing
Replace the 128 MB flat material_volume_buf (one u8 per voxel, all 4096
slots pre-allocated) with visibility-driven sparse storage using u16
material IDs. Only sub-blocks containing visible faces are stored,
packed contiguously in a shared buffer that grows on demand.

Chunk (src/chunk.rs):
- Widen material_block_ids() return type from [u8; 32768] to [u16; 32768]
- Remove truncating `as u8` cast, preserving full BlockId range

Build count shader (build_count.cs.hlsl):
- Compute 64-bit sub_mask via groupshared accumulator and InterlockedOr
- Each thread marks sub-blocks at (col/8, row/8, layer/8) with
  direction-specific coordinate mapping

Material pack shader (material_pack.cs.hlsl, NEW):
- Dispatch (popcount(sub_mask), 1, 1) workgroups of 256 threads
- Each workgroup copies one 1024-byte sub-block from staging to packed buf
- Reads dst_base and sub_mask from material_range_buf to stay within
  8-byte push constant limit

MaterialPackPipeline (build.rs):
- Bind group: staging (RO), material_buf (RW), material_range_buf (RO)
- 8-byte push constants: staging_index, slot

GPU world (world.rs):
- Remove material_volume_buf, insert() material param, write_material(),
  update_material()
- Add material_range_buf (64 KB), material_buf (16 MB initial, growable),
  material_staging_buf (4 MB transient), material_allocator
- dispatch_counts() accepts material_for callback, stages u16 material at
  dispatch time
- process_build_feedback() reads sub_mask from staging, allocates material
  ranges, dispatches material_pack passes
- grow_material_buf() doubles buffer on exhaustion: GPU-copies old data in
  the same encoder, rebuilds material_pack_bg and render_bg bind groups
- Stats track dynamic material_buf_used and material_buf_total

Render shaders:
- VS reads material_range_buf, passes mat_base + sub_mask as flat varyings
- PS resolves block ID via countbits(sub_mask & ((1 << sub_idx) - 1))
  plus aligned u16 load from packed material_buf
- Render bind group gains material_range_buf at binding 4, shifts
  remaining bindings 5-9

Chunk manager (chunk_manager.rs):
- Remove material extraction from insert path
- rebuild_dirty provides material_for closure reading from CPU World

+++ src/chunk.rs
@@ -183,13 +183,13 @@ impl Chunk {
 
     /// Build a per-voxel block ID array for GPU upload.
     ///
-    /// Each byte contains the raw [`BlockId`] value for the voxel at
+    /// Each element contains the raw [`BlockId`] value for the voxel at
     /// that position, resolved through the local palette. Unoccupied
     /// voxels are zero (air). Layout: `data[z * 1024 + y * 32 + x]`.
-    pub fn material_block_ids(&self) -> [u8; 32768] {
+    pub fn material_block_ids(&self) -> [u16; 32768] {
         let palette_data = self.material.as_slice();
         let occ_data     = self.occupancy.as_raw();
-        let mut out      = [0u8; 32768];
+        let mut out      = [0u16; 32768];
 
         for z in 0..32usize {
             for y in 0..32usize {
@@ -202,7 +202,7 @@ impl Chunk {
                 for x in 0..32usize {
                     if word & (1 << x) != 0 {
                         let idx = palette_data[row_base + x] as usize;
-                        out[row_base + x] = self.palette[idx].raw() as u8;
+                        out[row_base + x] = self.palette[idx].raw();
                     }
                 }
             }
@@ -340,8 +340,8 @@ mod tests {
         let ids = chunk.material_block_ids();
 
         // The array should contain raw BlockId values, not palette indices.
-        assert_eq!(ids[0], stone.raw() as u8);
-        assert_eq!(ids[1], dirt.raw() as u8);
+        assert_eq!(ids[0], stone.raw());
+        assert_eq!(ids[1], dirt.raw());
 
         // Unoccupied voxel should be 0 (air).
         assert_eq!(ids[2], 0);
